Move WithViewStore to file
